Fix resultsdb consumer for data dict items being lists
We noticed yesterday that the resultsdb consumer wasn't working
right in production - Bodhi was never updating gating status in
response to resultsdb.result.new messages, even when it should.
nirik and I figured out the problem: in resultsdb.result.new
messages, the values in the "data" dict are single-item lists
for some reason. The consumer code expected them to be strings,
as the equivalent values are in waiverdb.waiver.new messages.
The tests didn't catch this because I used the wrong format in
the sample messages used in the tests.

This updates `update_from_db_message` to handle the values being
lists, and makes the sample messages in the tests *be* lists to
better match the real-world case.

Signed-off-by: Adam Williamson <awilliam@redhat.com>
(cherry picked from commit 0599836515cb8792df11e257b70cb7c67f8c2356)