Build RPMs in ci
Add a new command to bodhi-ci to allow creation of RPMs and SRPMs. and
provide them back to the host if archive option is enabled.

Also, changes the integration tests for fedora releases (f35,f34, and
rawhide) to build the RPMs and install them in the container to test
from the RPMs.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>