Enhance UpdateReadyForTesting message to be more like the others
The UpdateReadyForTesting message is an odd duck. It was first
made much like all the other update messages, but then changed
to look completely different in response to the requirements of
the CI team:

https://pagure.io/fedora-ci/general/issue/70

If anyone had bothered to ask, I would've said that the QA team
would quite like for the message to look just like all the
others, since we'd already implemented openQA test scheduling
based on the other messages, and this custom format does not
include some information that is useful for that scheduling,
such as whether the update is in the critical path.

In an attempt to keep everyone happy, this replaces message V1
with V2, which inherits from V1 and adds the 'update' property
that all the other update messages have.

Signed-off-by: Adam Williamson <awilliam@redhat.com>