Migrate the client to use OIDC
The implementation uses the authlib integration for `requests`.

There are two modes for authentication, depending on the OIDC provider's
capability to do OOB (Out Of Band).

If the provider cannot do OOB, it will spawn a temporary HTTP server to
receive the redirect, so it only works if bodhi-client is running in the
same machine as the browser.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>