Prepare the client for the switch to OIDC
This change prepares the client to be compatible with an OIDC-enabled
server. It is also compatible with the current server, because
bodhi-server currently ignores the new `method` parameter.

We should widely deploy the client with this change before deploying the
OIDC-enabled server.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>
(cherry picked from commit 973aabd998b8820b3920dbae30718d186fe4b356)