Don't show duplicate test results on critpath updates (#4320)
Since b933a07 , we query on two Greenwave 'decision contexts'
for critical path updates. A query with a different decision
context but the same subjects will always return the same set
of test results. So we wind up showing every test result twice,
on the Automated Test Results tab, because we just send the
output from every Greenwave query that gets run through
`handle_results`.

To avoid this, let's skip `handle_results` when the decision
context ended in "_critpath". We know we would always also run
the non-critpath query, so we will get the results from that
query. We do *not* skip `handle_unsatisfied_requirements`,
because whether any requirements are unsatisfied can and will
differ between queries with the same subject(s) but different
decision contexts (which is why we run two queries in the first
place).

Signed-off-by: Adam Williamson <awilliam@redhat.com>