do the 3 koji builds concurrently
Previously, we were building the koji builds concurrently using the
github matrix, but when we were doing this, we were also building the
SRPMs individually for each job -- which was causing issues because we
encode the time in the RPMS version.

So to fix this, we removed the matrix, and sent the builds to koji
sequentially. This worked but was slow, as we had to wait for each koji
build to finish -- so it took about 12 minutes or so.

In this commit, we build the SRPMs in a seperate job, then upload them
as artifacts, then use them to kick off the koji builds sequentially.
Now building in koji takes about 3-4 minutes.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>