copy severity when obsoleting a security update
Previously, If a packager obsoletes a security update with a
non-security update type, the newer update is automatically
set to a security update. However, the severity of an update is
validated to be required on security updates.

This commit copies the severity of the obsoleted update to the new
update when changing the new update to a security update.

Resolves: #4441

Signed-off-by: Ryan Lerch <rlerch@redhat.com>