Fix critical path determination for Rawhide updates (#4480)
Currently Rawhide updates are never marked as critical path,
because we always use the release's `branch` as the `name` in
the PDC query, but for Rawhide this is incorrect. Bodhi's `branch`
for the release number that's currently Rawhide will be f37 (or
f38, f39... in future), but when querying PDC for the release
that's currently Rawhide, the `name` parameter must be "rawhide".
This fixes that, using the convention used elsewhere of the
release's `composed_by_bodhi` property indicating whether it's
Rawhide.

Signed-off-by: Adam Williamson <awilliam@redhat.com>