CLI: don't override flags for a release when editing it
Because the defaults in release_options for composed_by_bodhi and
create_automatic_updates were set to True and False respectively,
and `edit_release` uses those options, when you edit a release -
even if you don't specifically pass any argument to set either of
those flags - the CLI will change whatever the current values of
those flags are for the release you edit to those defaults. So
if you edit a release with composed_by_bodhi set to False and
don't pass `--composed-by-bodhi` or `--not-composed-by-bodhi`,
the flag will be changed to True.

This solves that by changing the defaults to None - which causes
`edit_release` to properly leave them at their current values
unless the user specifically passed a relevant argument - and
having `create_release` set them to the expected default value
if they come in as None.

We also throw away the explicit handling of composed_by_bodhi
that was added to both functions in 3de35881, because it is
unnecessary. It is already coming into both functions as a kwarg;
handling it like this changes nothing at all. Just adding the
option to release_options was all that was ever needed.

Signed-off-by: Adam Williamson <awilliam@redhat.com>