Web UI test result view: generalize 'flavor' discovery for openQA
We have rather more 'flavors' than just server and workstation;
some tests run on server, workstation and kde, so cannot be
properly distinguished if we're only tagging server and
workstation. This should be a general way to get a correct
'flavor' string for openQA results, and because we check for
"fedora.updates-" in the scenario string, it shouldn't do
anything weird for results from CI that don't follow the same
conventions.

Signed-off-by: Adam Williamson <awilliam@redhat.com>