Allow reading critpath lists from json files
This adds another valid setting for critpath.type, 'json', which
means to read the critpath information from JSON files. These
are expected to be produced by the `critpath.py` script from
the releng repo (which I recently updated to output this data).
An additional setting, `critpath.jsonpath`, tells Bodhi where
to find the files. The files must be named for each Release's
'branch' value, which is the name of the dist-git branch for
that release.

Ultimately this has two benefits: let us stop using PDC for this
(we can just have a cron job to run `critpath.py` nightly or
whatever), which is one step to not needing PDC any more and
avoids unnecessary network traffic and server load, and allow
us to do critpath stuff by group. The JSON data as output by
`critpath.py` is grouped. This commit doesn't take advantage of
that yet - it flattens the data into a single big list when
reading it - but follow-up commits will add additional code
paths to determine which groups a given update is critpath for.

Signed-off-by: Adam Williamson <awilliam@redhat.com>