Add critpath_groups to Update model
This new column is a nullable space-separated string listing
the critical path groups, if any, of which packages in the
update are members. The empty string means "no package in the
update is a member of any critpath group" and is equivalent to
critpath=False. A value of null/None means the configured
critpath type does not expose grouped information (i.e. it's
not critpath.type=json).

We add a method and a util function to support figuring this
out. This is kind of the long and humdrum way around, but it
keeps things clearer overall than making get_critpath_components
return different things for different types, for e.g.

Signed-off-by: Adam Williamson <awilliam@redhat.com>