Only cache critpath components when retrieving them from PDC
This moves the `@memoized` decorator to the function for getting
critpath from PDC. The effect of this is we will not cache
the critpath when retrieving it from a local source (JSON files
or configuration). There's no real point in re-reading it from
the configuration on every run of the function, but it also
doesn't hurt much - it's probably no more resource-intensive
than running through the `@memoized` decorator anyway. There is
a benefit to re-reading the JSON file on every run: it allows
us to update the critical path regularly without restarting
Bodhi. This function is called only when creating or editing
updates, so I don't think we're going to overwhelm servers with
disk reads or anything by doing this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>