Tweak gating status calculation to consider queued/running results
Per https://pagure.io/greenwave/pull-request/661#comment-183377 ,
if the test system submits QUEUED / RUNNING "results", we can
tell whether there is one for a "missing" result in the greenwave
output by checking for a result_id. If the result is "missing"
but we have a result_id for it, that means it's queued/running.
If it has no result_id, Greenwave really didn't find any results.

Using this we can tweak the gating status calculation a bit: if
we have queued/running results for all required tests, even if
it's been more than two hours since the update was edited, still
count the state as "waiting" not "failed".

Signed-off-by: Adam Williamson <awilliam@redhat.com>
(cherry picked from commit 93c0bba563e04ded8ffa578383fbca8a21b43352)