Web UI gating summary: account for 'running' tests
We can distinguish between results that are completely missing,
and cases where we know the test is queued or running (here, we
just flatten that to 'running' for simplicity). Take advantage
of this to generate a more useful and detailed gating status
summary, when the information is available.

This should also prevent us duplicating running test results in
the result table, which right now I believe we would do in any
case where queued/running test info is available. With this
change, we only create an ABSENT row for results that are truly
missing.

Signed-off-by: Adam Williamson <awilliam@redhat.com>
(cherry picked from commit 6eea927728a4e20d339b9c557f5f2dffb93e9b8c)