Handle Greenwave failed requirements to do with gating.yaml
Greenwave has several 'unsatisfied requirement' types that are
about issues with remote rule gating.yaml files, they don't
exactly map to 'failed tests'. Currently we handle these quite
badly; they are sort of counted as 'failed tests' in the
summary text, and not shown on the Automated Test Results tab
at all, so it's not at all clear what's going on if an update is
gated due to one of these problems.

This should handle them a bit better. They will no longer be
counted as "test results" in the summary (instead they get
counted separately as 'errors handling remote rules'), and they
will be shown on the Automated Test Results tab as errors with
notes that should give at least an idea of the problem.

Fixes: #5361

Signed-off-by: Adam Williamson <awilliam@redhat.com>