updates.html: include satisfied requirements in the table
I noticed today that tests which are 'required' aren't marked
as such in the Bodhi webUI if they passed. Only queued or failed
or missing tests are shown as 'required' (as shown by the
black asterisk with a mouseover comment, on the Automated Tests
tab). This is because we're not pushing satifised requirements
into the list that's used to track which tests are required.

Signed-off-by: Adam Williamson <awilliam@redhat.com>

end

(cherry picked from commit 36cc800257005041362be3a8e9cf431993a3084c)