Use `os.path.expanduser("~")` instead of `os.environ["HOME"]`
On openQA staging, some code that runs `bodhi updates download`
via `subprocess` was failing. Somehow in that context "HOME"
wasn't in `os.environ`, so bodhi would crash, because it
assumes that will be the case.

This should make it more robust, by using os.path.expanduser
instead. As documented:
https://docs.python.org/3/library/os.path.html#os.path.expanduser
it has fallback ways to get the path if the obvious one doesn't
work.

Signed-off-by: Adam Williamson <awilliam@redhat.com>