webui: add 'For help debugging' message for FCOS tests
We've added new `coreos.*` tests that now show up in the Bodhi UI for
some packages. These tests are not yet gating, but to be ready for when
they are, let's be nice and link to the #coreos Matrix channel where
packagers can get debugging help like for the other ones.

While we're here, add final periods to each sentence.

Signed-off-by: Jonathan Lebon <jonathan@jlebon.com>