Send UpdateReadyForTesting when update is edited and builds change
Test systems should re-test an update if it's edited and its
builds change. Before b33609db3b022afc114d014fde1b1c926f6d975b
we could rely on UpdateReadyForTesting being sent (after a delay)
when this happened, because editing the builds causes the update
to be sent back to pending state and then to testing state, and we
were publishing the message any time an update hit testing state.
But now we publish the message on update creation instead of when
the update reaches testing status, that is no longer the case. So
instead we should directly publish UpdateReadyForTesting when an
update is edited and the builds change.

Signed-off-by: Adam Williamson <awilliam@redhat.com>