Publish UpdateReadyForTesting whenever an update is created
See https://pagure.io/fedora-ci/general/issue/436 . This is the
behaviour test systems want, based on real-world experience.
We do not want test systems to wait for updates to reach the
updates-testing repo before they test; we want them to test as
soon as the update is created. All the test systems that exist
pull the builds to be tested from Koji, they do not rely on them
being present in updates-testing.

Waiting for the update to reach updates-testing before tests run
can mean a wait of up to 24 hours, which is obviously not ideal
when people want test results fast. It also means we can't gate
push to updates-testing on tests that trigger on this message.

openQA currently listens out for other messages to use as a
proxy for 'update was created' in order to mitigate this problem,
but Fedora CI does not do so, and consequently its tests often
trigger late. This change, along with the earlier change to
publish this message when an update's builds are changed, will
allow us to simplify openQA's logic considerably (it can just
trigger any time it sees this message) and improve CI's
scheduling considerably without any need to make changes to the
trigger logic on the CI side.

Signed-off-by: Adam Williamson <awilliam@redhat.com>
(cherry picked from commit d3765ed0908e21f8ee69d584edf662b4ffc58645)