Remove all taskotron/autoqa remnants (#2340) (#3938)
This removes all the code for querying taskotron, related
config settings, and the 'requirements' column from the Update
and Build models that backed it. As discussed in #2340 and #3938,
this whole mechanism is more or less obsoleted by greenwave, and
was rarely used (since 2017, only one person used it - in 2019 -
and we checked with them and they don't need it any more).

Signed-off-by: Adam Williamson <awilliam@redhat.com>
(cherry picked from commit 985ed593fa4949ccc4a4c100514b20220f7886a9)