Update type of User.email with recent Alembic
Alembic gives me this schema migration just from running current
Alembic (1.12.1) on current develop. The type of this column
in the Python code hasn't changed for years, so I assume it's
a change in Alembic. I don't want this change in the migration
I'm planning, so committing it separately first.

Signed-off-by: Adam Williamson <awilliam@redhat.com>