Always disable autopush on negative karma, regardless of status
Only disabling autopush for updates in 'testing' status makes no
sense. @trishnaguha explained in
https://github.com/fedora-infra/bodhi/pull/1556#discussion_r123996802
that their rationale was that updates in 'pending' status should
get obsoleted (instead?), but this...just isn't right. We only
obsolete updates when they reach the negative karma threshold,
which is usually much stricter than just a single -1 (by default
it's a karma *total* of -3). We want to disable autopush for any
update as soon as it gets a single -1, regardless of whether it's
in pending or testing status, and regardless of whether we might
later obsolete it.

Signed-off-by: Adam Williamson <awilliam@redhat.com>