Drop a bogus branch from check_karma_thresholds
This has been hanging around for a while, but it's fairly
useless. At some point this is where we were planning to post
the testing_approval_msg message if the update reached the
critpath.min_karma threshold, but that is actually done by
the approve_testing.py script nowadays. All we do here is log a
message, and it's actually a bogus message anyway, because
stable_karma is *not* the manual push threshold. stable_karma
is the autopush threshold; it's really meaningless if autokarma
is not set. So it doesn't make any sense to log this message
when we hit stable_karma but autopush is not set. We should just
do nothing on that path.

Note this slightly changes when we set date_approved, but the
current behaviour around date_approved makes no sense anyway,
so the next commit revamps it entirely to make much more sense.

Signed-off-by: Adam Williamson <awilliam@redhat.com>