Add BCD: a standalone, container-based dev environment
The current development environment requires four separate VMs
(three from tinystage and one for Bodhi itself), plus some
containers running inside the Bodhi VM. It's pretty heavy!

This provides an alternative development environment that is
standalone and entirely based on containers: a postgres
container, a waiverdb container, a greenwave container, a
rabbitmq container, an ipsilon container, and a bodhi container
that functions similarly to the bodhi VM from the existing
environment. There is no FreeIPA or FAS backing the ipsilon
instance, we just use ipsilon's testauth mode, with some
configuration to allow for testing different group memberships.

The containers are Podman containers orchestrated by an ansible
playbook, around which the `bcd` command is a light wrapper.
They share a network namespace via a pod, and everything is
accessed from the host as `localhost`, without SSL or Apache
(these could be added if we really want, but it seemed
unnecessary).

Signed-off-by: Adam Williamson <awilliam@redhat.com>