Fix Automated Tests table rows for absent and error results
This fixes my own oversight from
https://github.com/fedora-infra/bodhi/pull/5480 . In changing
the signature of make_row there, I did not remember to update
all these other calls to it. This means make_row goes a bit
crazy trying to treat 'FAILED' or 'ABSENT' as a subject and
an error string or empty string as the outcome, and winds up
rendering what looks to the user like an empty row. This should
fix it so we actually get rows with sensible content again when
a result is absent or there's some kind of remote rule error.
Thanks to @decathorpe for finding this.

Signed-off-by: Adam Williamson <awilliam@redhat.com>