Rewrite test_approve_testing.py to match current implementation
On close inspection, this test suite has become a little weird
over the years.

When it was introduced and most of it was written, approve_testing
did somewhat more than it does now. It handled autokarma push
(which is now handled by Update.check_karma_thresholds), and it
carried much more of the logic for deciding if updates met the
relevant requirements or not (most of which is now handled by
Update.meets_testing_requirements; all that's left in
approve_testing is the check of whether the update meets the
autopush-for-time-in-testing threshold).

A lot of the tests are really checking those things - all the
minute variations of "is the update critical path or not?",
"does the update have autokarma enabled?", various time spent
in testing and karma levels are really about testing that stuff.
It doesn't make sense to test that here any more. Those things
should be tested elsewhere, mainly in the tests of
meets_testing_requirements. At least one of these tests -
test_autotime_update_with_autokarma_met_karma_and_time_requirements_get_pushed -
actually still passes if you completely remove the call to
approve_testing_main()!

This completely rewrites the test suite to focus on testing what
approve_testing actually does now. To do this I evaluated every
existing test, and also evaluated the current code of approve_
testing to try and make sure I was testing every path through it.
I'll include those notes in the pull request as they're long.
Some of the existing tests that are still useful are kept, but
tweaked slightly to take advantage of the new setup_method and
assertion helpers, to reduce duplication.

Signed-off-by: Adam Williamson <awilliam@redhat.com>