Refactor all tests for check_karma_thresholds
The existing tests that ultimately test the stuff handled
by check_karma_thresholds are spread around, both within
individual test files and across them. There's kind of a grab
bag of tests of different paths that have been added a bit
haphazardly over the years.

This tries to reconcile them all into a set of comprehensive
and consistent tests that all live in one place. I did quite
a lot of checking to try and make sure everything that was
currently tested still is, and that some things which weren't
currently tested are now. This is a precursor to changing
some of the behaviour of check_karma_thresholds in later
commits.

Signed-off-by: Adam Williamson <awilliam@redhat.com>