Fix bodhi-shell in the BCD development environment
While working on removing Vagrant I noticed this didn't work, so
let's fix it. It runs into the same shebang issue I had problems
with for a couple of other things, and it turns out to be easiest
to just forget about the 'normal' implementation and replace it
with a simple alias, since we know where the config file is in
the development environment.

Signed-off-by: Adam Williamson <awilliam@redhat.com>