client `updates download`: get signed packages when possible
This tries to get signed packages when possible, by calling gpg
on the key file in /etc/pki/rpm-gpg to figure out the key ID,
and passing it to Koji. If we can't work it out - because the
key file isn't there on this system, or gpg isn't there, or
whatever - we just fall through to getting unsigned packages.
There's also a `--no-gpg` to disable this behaviour.

Signed-off-by: Adam Williamson <awilliam@redhat.com>