client `updates download`: get key from dist-git if no file
This extends the 'try and find the correct signing key ID and
get signed packages' mechanism to try and find the key file from
fedora-repos dist-git if we can't find it as a local file. This
will make it work on non-Fedora systems (and Fedora systems where
we can't find the key file for some reason), but makes it a bit
more complicated, and will need updating when dist-git moves.

Signed-off-by: Adam Williamson <awilliam@redhat.com>