Fix indicating waived failures when there is no scenario
On the Automated Tests tab, when a required test failed but the
failure has been waived, we intend to show a thumbs-up icon with
a tooltip with details about the waiver. However, if the failed
test has no 'scenario', this doesn't work, because there's a
hash key mismatch (we store the waiver under key "testcase##null"
but look it up under key "testcase##no_scenario").

This should fix that, and mean we correctly note waived failures
for Fedora CI results (openQA ones always have a scenario).

Signed-off-by: Adam Williamson <awilliam@redhat.com>