Show waived missing results in the test result table
We generate the rows for missing results based on the requirements
in the greenwave response, because they do not show up in the
'results' (obviously, because they're...missing). However, we
were only doing this for ones in unsatisfied_requirements. Waived
missing results show up in satisfied_requirements, though. So
when a missing result was waived, it didn't show up in the result
table at all.

This fixes that by looking for missing-waived requirements when
we go through satisfied_requirements and handling them the same
way. We also have to tweak handle_results a bit to use
result.subject_identifier instead of result.item, because
-waived requirements (TestResultWaived class in greenwave) don't
have the item dict, which is considered 'legacy' in greenwave. I
also corrected the comment here - we're dealing with completely
different data structures when we handle these missing results
versus when we handle actual passed-through results.

Signed-off-by: Adam Williamson <awilliam@redhat.com>