Automated Tests table: more accurate indication of required tests (#5932)
The tracking of which tests are "required" was too naive; it was
only going by test case name. That meant that if the same test
was required for one package but not another, it would be shown
as required for both. It also meant that if the test was required
in one scenario but not another (e.g. on a different image, or a
different arch), it would be shown as required for all.

This adds the subject identifier and the scenario to the
identifier for each "required" test.

Signed-off-by: Adam Williamson <awilliam@redhat.com>